+++ tango/step.py
@@ -117,8 +117,9 @@ class Step(Registrable, Generic[T]):
       of inputs.
     :param step_resources: gives you a way to set the minimum compute resources required
       to run this step. Certain executors require this information.
-    :param workspace_metadata: use this to specify additional metadata for your workspace.
-      The type of metadata depends on your workspace.
+    :param step_metadata: use this to specify additional metadata for your step.
+      This is added to the :attr:`METADATA` class variable to form the ``self.metadata`` attribute.
+      Values in ``step_metadata`` take precedence over ``METADATA``.
 
     .. important::
         Overriding the unique id means that the step will always map to this value, regardless of the inputs,
@@ -157,6 +158,11 @@ class Step(Registrable, Generic[T]):
     the model output, not about how many outputs you can produce at the same time.
     """
 
+    METADATA: Dict[str, Any] = {}
+    """
+    Arbitrary metadata about the step.
+    """
+
     _UNIQUE_ID_SUFFIX: Optional[str] = None
     """
     Used internally for testing.
@@ -170,7 +176,7 @@ class Step(Registrable, Generic[T]):
         step_config: Optional[Dict[str, Any]] = None,
         step_unique_id_override: Optional[str] = None,
         step_resources: Optional[StepResources] = None,
-        workspace_metadata: Optional[Dict[str, Any]] = None,
+        step_metadata: Optional[Dict[str, Any]] = None,
         **kwargs,
     ):
         if self.VERSION is not None:
@@ -245,7 +251,9 @@ class Step(Registrable, Generic[T]):
         ] = None  # This is set only while the run() method runs.
         self._config = step_config
         self.step_resources = step_resources
-        self.workspace_metadata = workspace_metadata
+        self.metadata = deepcopy(self.METADATA)
+        if step_metadata:
+            self.metadata.update()
 
     @classmethod
     def massage_kwargs(cls, kwargs: Dict[str, Any]) -> Dict[str, Any]: