+++ tango/step_info.py
@@ -228,9 +228,9 @@ class StepInfo(FromParams):
     The raw config of the step.
     """
 
-    workspace_metadata: Optional[Dict[str, Any]] = None
+    metadata: Optional[Dict[str, Any]] = None
     """
-    Workspace metadata from the step. This comes from the ``workspace_metadata``
+    Metadata from the step. This comes from the ``step_metadata``
     argument to the :class:`~tango.step.Step` class.
     """
 
@@ -326,7 +326,7 @@ class StepInfo(FromParams):
             dependencies={dep.unique_id for dep in step.dependencies},
             cacheable=step.cache_results,
             config=replace_steps_with_unique_id(config),
-            workspace_metadata=step.workspace_metadata,
+            metadata=step.metadata,
             **kwargs,
         )